+++ collective/portlet/calendar/calendar.py
@@ -266,12 +266,16 @@ class Renderer(base.Renderer):
         if index == 'start':
             last_day = self.calendar._getCalendar().monthrange(year, month)[1]
             calendar_date = self.calendar.getBeginAndEndTimes(last_day, month, year)[1]
-            criteria['query'].append(calendar_date)
+            # add the last day in the month only if smaller than collection's date
+            if calendar_date<max(criteria['query']) or criteria['range'] != 'max':
+                criteria['query'].append(calendar_date)
             if criteria['range'] == 'min':
                 criteria['range'] = 'minmax'
         elif index == 'end':
             calendar_date = self.calendar.getBeginAndEndTimes(1, month, year)[0]
-            criteria['query'].append(calendar_date)
+            # add the fist day in the month only if greater than collection's date
+            if calendar_date>min(criteria['query']) or criteria['range'] != 'min':
+                criteria['query'].append(calendar_date)
             if criteria['range'] == 'max':
                 criteria['range'] = 'minmax'
         self.options[index] = criteria